Conducted several SVN updates. Currently at Runmodains: ver. 143 And RunUO-SVN: ver 663 Using UO 7.0.12.0 Unable to go higher atm cause of conflict with code. An easy check is you Will loose text box menu with NPCs when you single click on them.
Had directory modifications and did full check of empty directory's.


Main updated I preformed were:

Created spawns.map and hand.map due to SVNs moding premium spawner. Crashed spawner and had to create to redo spawn world.
Also because of mod with spawner, unable to make them player sensitive.

Fixed to have BonusMaxWeight on char and Misc items pertaining to it. Moded the core files along with container.

Tried to make ToolBar for GMs to auto re spawn if GM died - Didn't work.

Still testing but think the Stable Post is working correctly and pets wont disappear. ** Still testing**

Changed some of the basic gumps, and coloring of text.

Made Dump command text show only to player and not whole area.

Fixed srink issue with Evo so it now doesn't show a little yellow ball. Along with other mobs that were doing it also.

Fixed Evo max evolve to stage 7 and also changed graph of different evolutions.

Added pet dye reset potion to the pet dye vendor.

Adjusted some of the reg seeds to try and make so that will disappear if not on dirt tile. But problem is code only supported X, Y check and not Z.

Added David Banner, incredible hulk - Attacks, and changes when you kill David to hulk with delete timer so wont have multi spawns of hulk.

**Note: didn't upload save data cause of local backups.

+++ Scripts/Customs/PERSONALY ADDED/ACTIVE/Items/LinkedTeleportersBag/LinkedGate.cs
@@ -64,7 +64,7 @@ namespace Server.Lokai
 			m_Target = MateGate.Location;
 			m_TargetMap = MateGate.Map;
 			
-			int flags = m.NetState == null ? 0 : m.NetState.Flags;
+			int flags = m.NetState == null ? 0 : (int)m.NetState.Flags;
 
 			if ( Factions.Sigil.ExistsOn( m ) )
 			{
